# Remove 1 allocation from point.GeoHash