--- a/benchmarks_point_test.go
+++ b/benchmarks_point_test.go
@@ -10,6 +10,7 @@ func BenchmarkPointDistanceFrom(b *testing.B) {
 	p1 := geo.NewPoint(-122.4167, 37.7833)
 	p2 := geo.NewPoint(37.7833, -122.4167)
 
+	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		p1.DistanceFrom(p2)
@@ -20,6 +21,7 @@ func BenchmarkPointSquaredDistanceFrom(b *testing.B) {
 	p1 := geo.NewPoint(-122.4167, 37.7833)
 	p2 := geo.NewPoint(37.7833, -122.4167)
 
+	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		p1.SquaredDistanceFrom(p2)
@@ -29,6 +31,7 @@ func BenchmarkPointSquaredDistanceFrom(b *testing.B) {
 func BenchmarkPointQuadKey(b *testing.B) {
 	p := geo.NewPoint(-122.4167, 37.7833)
 
+	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		p.Quadkey(60)
@@ -38,6 +41,7 @@ func BenchmarkPointQuadKey(b *testing.B) {
 func BenchmarkPointQuadKeyString(b *testing.B) {
 	p := geo.NewPoint(-122.4167, 37.7833)
 
+	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		p.QuadkeyString(60)
@@ -47,6 +51,7 @@ func BenchmarkPointQuadKeyString(b *testing.B) {
 func BenchmarkPointGeoHash(b *testing.B) {
 	p := geo.NewPoint(-122.4167, 37.7833)
 
+	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		p.GeoHash()
@@ -56,6 +61,7 @@ func BenchmarkPointGeoHash(b *testing.B) {
 func BenchmarkPointGeoHashInt64(b *testing.B) {
 	p := geo.NewPoint(-122.4167, 37.7833)
 
+	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		p.GeoHashInt64(60)
@@ -65,6 +71,7 @@ func BenchmarkPointGeoHashInt64(b *testing.B) {
 func BenchmarkPointNormalize(b *testing.B) {
 	p := geo.NewPoint(5, 6)
 
+	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		p.Normalize()
@@ -75,6 +82,7 @@ func BenchmarkPointEquals(b *testing.B) {
 	p1 := geo.NewPoint(5, 6)
 	p2 := geo.NewPoint(5, 7)
 
+	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		p1.Equals(p2)
@@ -84,6 +92,7 @@ func BenchmarkPointEquals(b *testing.B) {
 func BenchmarkPointClone(b *testing.B) {
 	p := geo.NewPoint(5, 6)
 
+	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		p.Clone()