```
benchmark                      old ns/op     new ns/op     delta
BenchmarkPointGeoHash          432           309           -28.47%

benchmark                      old allocs     new allocs     delta
BenchmarkPointGeoHash          2              1              -50.00%

benchmark                      old bytes     new bytes     delta
BenchmarkPointGeoHash          32            16            -50.00%
```

FYI @mlerner