@mlerner 2015-02-20
:+1: for speed